+++ bids/modeling/hrf.py
@@ -1,5 +1,5 @@
 """
-This module is for hemodynamic reponse function (hrf) specification.
+This module is for hemodynamic response function (hrf) specification.
 Here we provide for SPM, Glover hrfs and finite timpulse response (FIR) models.
 This module closely follows SPM implementation
 
@@ -43,7 +43,7 @@ def _gamma_difference_hrf(tr, oversampling=50, time_length=32., onset=0.,
     """
     from scipy.stats import gamma
     dt = tr / oversampling
-    time_stamps = np.linspace(0, time_length, np.rint(float(time_length) / dt).astype(np.int))
+    time_stamps = np.linspace(0, time_length, np.rint(float(time_length) / dt).astype(int))
     time_stamps -= onset
     hrf = gamma.pdf(time_stamps, delay / dispersion, dt / dispersion) -\
         ratio * gamma.pdf(
@@ -239,7 +239,7 @@ def _sample_condition(exp_condition, frame_times, oversampling=50,
 
     hr_frame_times = np.linspace(frame_times.min() + min_onset,
                                  frame_times.max() * (1 + 1. / (n - 1)),
-                                 np.rint(n_hr).astype(np.int))
+                                 np.rint(n_hr).astype(int))
 
     # Get the condition information
     onsets, durations, values = tuple(map(np.asanyarray, exp_condition))
@@ -250,7 +250,7 @@ def _sample_condition(exp_condition, frame_times, oversampling=50,
 
     # Set up the regressor timecourse
     tmax = len(hr_frame_times)
-    regressor = np.zeros_like(hr_frame_times).astype(np.float)
+    regressor = np.zeros_like(hr_frame_times).astype(float)
     t_onset = np.minimum(np.searchsorted(hr_frame_times, onsets), tmax - 1)
     regressor[t_onset] += values
     t_offset = np.minimum(
@@ -417,7 +417,7 @@ def compute_regressor(exp_condition, hrf_model, frame_times, con_id='cond',
     oversampling : int, optional
         oversampling factor to perform the convolution
     fir_delays : 1D-array-like, optional
-        delays (in seconds) used in case of a finite impulse reponse model
+        delays (in seconds) used in case of a finite impulse response model
     min_onset : float, optional
         minimal onset relative to frame_times[0] (in seconds)
         events that start before frame_times[0] + min_onset are not considered